Feature 178 wip second part
