--- resilience4j-spring-boot/src/main/java/io/github/resilience4j/ratelimiter/annotation/RateLimiter.java
@@ -1,40 +1,0 @@
-/*
- * Copyright 2017 Bohdan Storozhuk
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.github.resilience4j.ratelimiter.annotation;
-
-import java.lang.annotation.Documented;
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-/**
- * This annotation can be applied to a class or a specific method.
- * Applying it on a class is equivalent to applying it on all its public methods.
- * The annotation enables throttling for all methods where it is applied.
- * Throttling monitoring is performed via a rate limiter.
- * See {@link io.github.resilience4j.ratelimiter.RateLimiter} for details.
- */
-@Retention(value = RetentionPolicy.RUNTIME)
-@Target(value = {ElementType.METHOD, ElementType.TYPE})
-@Documented
-public @interface RateLimiter {
-    /**
-     * Name of the rate limiter
-     * @return
-     */
-    String name();
-}
