+++ resilience4j-spring/build.gradle
@@ -1,0 +1,13 @@
+dependencies {
+    compile ( libraries.aspectj )
+    compileOnly ( libraries.hibernate_validator )
+    compileOnly ( libraries.spring_4_core, libraries.spring_4_context )
+    compile project(':resilience4j-annotations')
+    compile project(':resilience4j-consumer')
+    compile project(':resilience4j-circuitbreaker')
+    compile project(':resilience4j-ratelimiter')
+    compileOnly project(':resilience4j-prometheus')
+    compileOnly project(':resilience4j-metrics')
+    testCompile project(':resilience4j-prometheus')
+    testCompile project(':resilience4j-metrics')
+}
